requirements.txt
evennia==5.0.1
psycopg2-binary
dj-database-url
openai
requests
resend
django-cors-headers
